Generated remote interfaces now extend the parameter interfaces, throwing checke d exceptions on every generated method, and so throwing RemoteException.
StubbingInvocationHandler now alsa catches RemoteException, that may happen on r
miStub.checkConnection(), and marks the connectionFailure on rmiStub.
When invoking a method, StubbingInvocationHandler checks that those methods are
not inherited from Object. If that's the case, it uses its own methods, and not
the remote ones.

StubbingInvocationHandler.equals(Object) can tell it is equal to a Proxy, if tha
t proxy wraps another StubbingInvocationHandler, and that invocationHandler equa
ls it.



git-svn-id: http://flaptor-util.googlecode.com/svn/trunk@148 a957e277-6143-0410-b0ff-25ec9014f3a9